requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
openai>=1.12.0
python-dotenv>=1.0.0
python-docx>=1.1.0
